requirements.txt
fastapi
uvicorn
sqlalchemy
packaging
ipaddress
python-jose[cryptography]
python-dotenv
